refactor: 설정 파일 구조 재편 (src/lib → src/config, src/utils)
- src/lib/auth.ts → src/config/auth.ts
- src/lib/prisma.ts → src/config/prisma.ts
- src/lib/swagger/ → src/config/swagger/
- src/lib/utils.ts → src/utils/index.ts
- 모든 @/lib/utils 임포트 경로를 @/utils로 변경
- @/lib/auth, @/lib/swagger 경로도 각각 @/config/auth, @/config/swagger로 변경
