+++ frontend/src/app/(app)/daily/_hooks/use-hovered-time.tsx
@@ -1,6 +1,6 @@
 import { useCallback, useMemo, useState } from 'react'
 
-import { dateToMinutes, minutesToDayjs } from '@/lib/utils'
+import { dateToMinutes, minutesToDayjs } from '@/utils'
 import { useDateStore } from '@/store'
 import type { Minutes } from '@/types'
 import type { Execution } from '@/types/execution'
