+++ frontend/src/app/(app)/weekly/_hooks/use-hovered-time.ts
@@ -1,6 +1,6 @@
 import { useCallback, useMemo, useState } from 'react'
 
-import { dateToMinutes, minutesToDayjs } from '@/lib/utils'
+import { dateToMinutes, minutesToDayjs } from '@/utils'
 import { useDateStore } from '@/store'
 import type { Minutes } from '@/types'
 import type { Plan } from '@/types/plan'
