+++ frontend/src/app/(app)/weekly/_utils/index.ts
@@ -1,7 +1,7 @@
 import dayjs from 'dayjs'
 
 import { HOURS_PER_DAY, MINUTES_PER_HOUR } from '@/constants'
-import { dateToMinutes } from '@/lib/utils'
+import { dateToMinutes } from '@/utils'
 import type { Minutes } from '@/types'
 import type { Plan } from '@/types/plan'
 import { DAYS_COUNT, ROW_HEIGHT } from '~/weekly/_constants'
