+++ frontend/src/components/ui/avatar.tsx
@@ -3,7 +3,7 @@
 import { Avatar as AvatarPrimitive } from 'radix-ui'
 import * as React from 'react'
 
-import { cn } from '@/lib/utils'
+import { cn } from '@/utils'
 
 function Avatar({
   className,
