+++ frontend/src/components/ui/calendar.tsx
@@ -6,7 +6,7 @@ import * as React from 'react'
 import { type DayButton, DayPicker, getDefaultClassNames } from 'react-day-picker'
 
 import { Button, buttonVariants } from '@/components/ui/button'
-import { cn } from '@/lib/utils'
+import { cn } from '@/utils'
 
 function Calendar({
   className,
