+++ frontend/src/components/ui/color-picker.tsx
@@ -2,7 +2,7 @@
 
 import { CheckIcon } from 'lucide-react'
 
-import { cn } from '@/lib/utils'
+import { cn } from '@/utils'
 
 interface ColorPickerProps {
   colors: string[]
