+++ frontend/src/components/ui/input-group.tsx
@@ -6,7 +6,7 @@ import * as React from 'react'
 import { Button } from '@/components/ui/button'
 import { Input } from '@/components/ui/input'
 import { Textarea } from '@/components/ui/textarea'
-import { cn } from '@/lib/utils'
+import { cn } from '@/utils'
 
 function InputGroup({ className, ...props }: React.ComponentProps<'div'>) {
   return (
