+++ frontend/src/components/ui/label.tsx
@@ -3,7 +3,7 @@
 import { Label as LabelPrimitive } from 'radix-ui'
 import * as React from 'react'
 
-import { cn } from '@/lib/utils'
+import { cn } from '@/utils'
 
 function Label({ className, ...props }: React.ComponentProps<typeof LabelPrimitive.Root>) {
   return (
