+++ frontend/src/components/ui/sidebar.tsx
@@ -21,7 +21,7 @@ import {
   TooltipTrigger,
 } from '@/components/ui'
 import { useIsMobile } from '@/hooks'
-import { cn } from '@/lib/utils'
+import { cn } from '@/utils'
 
 const SIDEBAR_COOKIE_NAME = 'sidebar_state'
 const SIDEBAR_COOKIE_MAX_AGE = 60 * 60 * 24 * 7
