+++ frontend/src/components/ui/switch.tsx
@@ -3,7 +3,7 @@
 import { Switch as SwitchPrimitive } from 'radix-ui'
 import * as React from 'react'
 
-import { cn } from '@/lib/utils'
+import { cn } from '@/utils'
 
 function Switch({
   className,
