+++ frontend/src/components/ui/textarea.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react'
 
-import { cn } from '@/lib/utils'
+import { cn } from '@/utils'
 
 function Textarea({ className, ...props }: React.ComponentProps<'textarea'>) {
   return (
