+++ frontend/src/components/ui/tooltip.tsx
@@ -3,7 +3,7 @@
 import * as TooltipPrimitive from '@radix-ui/react-tooltip'
 import * as React from 'react'
 
-import { cn } from '@/lib/utils'
+import { cn } from '@/utils'
 
 function TooltipProvider({
   delayDuration = 0,
