--- frontend/src/lib/auth.ts
@@ -1,9 +1,0 @@
-import NextAuth from 'next-auth'
-import GitHub from 'next-auth/providers/github'
-import Google from 'next-auth/providers/google'
-
-export const { handlers, signIn, signOut, auth } = NextAuth({
-  providers: [Google, GitHub],
-  session: { strategy: 'jwt' },
-  pages: { signIn: '/login' },
-})
